@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "ampel-alerts"
-version = "0.8.2-alpha.10"
+version = "0.8.3-alpha.8"
 description = "Alert support for the Ampel system"
 authors = ["Valery Brinnel"]
 maintainers = ["Jakob van Santen <jakob.van.santen@desy.de>"]
@@ -26,19 +26,19 @@ include = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.9,<3.10"
-Sphinx = {version = ">=3.5.1,<5.0.0", optional = true}
+python = ">=3.10,<3.11"
+Sphinx = {version = ">=5.1.1,<5.2.0", optional = true}
 sphinx-autodoc-typehints = {version = "^1.11.1", optional = true}
-tomlkit = {version = "^0.7.0"}
-ampel-interface = {version = ">=0.8.2-alpha.7,<0.8.3"}
-ampel-core = {version = ">=0.8.2-alpha.7,<0.8.3"}
+tomlkit = {version = "^0.11.0", optional=true}
+ampel-interface = {version = "^0.8.3-alpha.15"}
+ampel-core = {version = "^0.8.3-alpha.20"}
 
 [tool.poetry.dev-dependencies]
-pytest = "^6.2.5"
+pytest = "^7.1.2"
 pytest-cov = "^3.0.0"
-pytest-mock = "^3.6.1"
-mongomock = "^3.23.0"
-mypy = "^0.910"
+pytest-mock = "^3.8.2"
+mongomock = "^4.1.2"
+mypy = "^0.930"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
