@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# File              : Ampel-alerts/setup.py
-# License           : BSD-3-Clause
-# Author            : vb <vbrinnel@physik.hu-berlin.de>
-# Date              : Unspecified
-# Last Modified Date: 15.12.2021
-# Last Modified By  : vb <vbrinnel@physik.hu-berlin.de>
+# File:                Ampel-alerts/setup.py
+# License:             BSD-3-Clause
+# Author:              valery brinnel <firstname.lastname@gmail.com>
+# Date:                Unspecified
+# Last Modified Date:  10.07.2022
+# Last Modified By:    valery brinnel <firstname.lastname@gmail.com>
 
 from setuptools import setup, find_namespace_packages
 
@@ -19,13 +19,13 @@ package_data = {
 
 setup(
     name = 'ampel-alerts',
-    version = '0.8.2.alpha-3',
-    description = 'Asynchronous and Modular Platform with Execution Layers',
+    version = '0.8.3.alpha-8',
+    description = 'Alice in Modular Provenance-Enabled Land',
     author = 'Valery Brinnel',
     maintainer = 'Jakob van Santen',
     maintainer_email = 'jakob.van.santen@desy.de',
     url = 'https://ampelproject.github.io',
     packages = find_namespace_packages(),
     package_data = package_data,
-    python_requires = '>=3.9,<4.0'
+    python_requires = '>=3.10,<3.11'
 )
