@@ -1,13 +1,12 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# File              : Ampel-alerts/ampel/abstract/AbsAlertFilter.py
-# License           : BSD-3-Clause
-# Author            : vb <vbrinnel@physik.hu-berlin.de>
-# Date              : 14.12.2017
-# Last Modified Date: 24.11.2021
-# Last Modified By  : vb <vbrinnel@physik.hu-berlin.de>
+# File:                Ampel-alerts/ampel/abstract/AbsAlertFilter.py
+# License:             BSD-3-Clause
+# Author:              valery brinnel <firstname.lastname@gmail.com>
+# Date:                14.12.2017
+# Last Modified Date:  24.11.2021
+# Last Modified By:    valery brinnel <firstname.lastname@gmail.com>
 
-from typing import Optional, Union
 from ampel.base.AmpelABC import AmpelABC
 from ampel.base.decorator import abstractmethod
 from ampel.base.LogicalUnit import LogicalUnit
@@ -18,7 +17,7 @@ class AbsAlertFilter(AmpelABC, LogicalUnit, abstract=True):
 	""" Base class for T0 alert filters """
 
 	@abstractmethod
-	def process(self, alert: AmpelAlertProtocol) -> Optional[Union[bool, int]]:
+	def process(self, alert: AmpelAlertProtocol) -> None | bool | int:
 		"""
 		Filters an alert.
 		
