@@ -1,13 +1,13 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# File              : Ampel-alerts/ampel/alert/FilteringAlertSupplier.py
-# License           : BSD-3-Clause
-# Author            : vb <vbrinnel@physik.hu-berlin.de>
-# Date              : 28.05.2020
-# Last Modified Date: 24.11.2021
-# Last Modified By  : vb <vbrinnel@physik.hu-berlin.de>
+# File:                Ampel-alerts/ampel/alert/FilteringAlertSupplier.py
+# License:             BSD-3-Clause
+# Author:              valery brinnel <firstname.lastname@gmail.com>
+# Date:                28.05.2020
+# Last Modified Date:  24.11.2021
+# Last Modified By:    valery brinnel <firstname.lastname@gmail.com>
 
-from typing import List, Iterator
+from typing import Iterator
 from ampel.abstract.AbsAlertSupplier import AbsAlertSupplier
 from ampel.model.UnitModel import UnitModel
 from ampel.log.AmpelLogger import AmpelLogger
@@ -31,7 +31,7 @@ class FilteringAlertSupplier(AbsAlertSupplier):
 	"""
 
 	supplier: UnitModel
-	match_ids: List[int]
+	match_ids: list[int]
 
 
 	def __init__(self, **kwargs) -> None:
