@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# File              : Ampel-alerts/ampel/alert/load/FileAlertLoader.py
-# License           : BSD-3-Clause
-# Author            : vb <vbrinnel@physik.hu-berlin.de>
-# Date              : 30.04.2018
-# Last Modified Date: 11.08.2021
-# Last Modified By  : vb <vbrinnel@physik.hu-berlin.de>
+# File:                Ampel-alerts/ampel/alert/load/FileAlertLoader.py
+# License:             BSD-3-Clause
+# Author:              valery brinnel <firstname.lastname@gmail.com>
+# Date:                30.04.2018
+# Last Modified Date:  11.08.2021
+# Last Modified By:    valery brinnel <firstname.lastname@gmail.com>
 
 from io import BytesIO
 from typing import List
@@ -18,7 +18,7 @@ class FileAlertLoader(AbsAlertLoader[BytesIO]):
 	"""
 
 	#: paths to files to load
-	files: List[str]
+	files: list[str]
 
 	def __init__(self, **kwargs) -> None:
 	
