@@ -1,14 +1,15 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# File              : Ampel-alerts/ampel/alert/reject/GeneralActiveAlertRegister.py
-# License           : BSD-3-Clause
-# Author            : vb <vbrinnel@physik.hu-berlin.de>
-# Date              : 26.05.2020
-# Last Modified Date: 24.11.2021
-# Last Modified By  : vb <vbrinnel@physik.hu-berlin.de>
+# File:                Ampel-alerts/ampel/alert/reject/GeneralActiveAlertRegister.py
+# License:             BSD-3-Clause
+# Author:              valery brinnel <firstname.lastname@gmail.com>
+# Date:                26.05.2020
+# Last Modified Date:  24.11.2021
+# Last Modified By:    valery brinnel <firstname.lastname@gmail.com>
 
 from struct import pack
-from typing import Optional, ClassVar, Tuple, Sequence, Union
+from typing import ClassVar
+from collections.abc import Sequence
 from ampel.protocol.AmpelAlertProtocol import AmpelAlertProtocol
 from ampel.alert.reject.GeneralAlertRegister import GeneralAlertRegister
 
@@ -16,9 +17,9 @@ from ampel.alert.reject.GeneralAlertRegister import GeneralAlertRegister
 class GeneralActiveAlertRegister(GeneralAlertRegister):
 	""" Logs: alert_id, stock_id, filter_res """
 
-	__slots__: ClassVar[Tuple[str, ...]] = '_write', 'alert_max', 'alert_min', 'stock_max', 'stock_min' # type: ignore
+	__slots__: ClassVar[tuple[str, ...]] = '_write', 'alert_max', 'alert_min', 'stock_max', 'stock_min' # type: ignore
 	_slot_defaults = {'alert_max': 0, 'alert_min': 2**64, 'stock_max': 0, 'stock_min': 2**64}
-	new_header_size: Union[int, str] = "+1000"
+	new_header_size: int | str = "+1000"
 
 	header_bounds: ClassVar[Sequence[str]] = ('alert', 'stock') # type: ignore
 	alert_min: int
@@ -26,7 +27,7 @@ class GeneralActiveAlertRegister(GeneralAlertRegister):
 	stock_min: int
 	stock_max: int
 
-	def file(self, alert: AmpelAlertProtocol, filter_res: Optional[int] = None) -> None:
+	def file(self, alert: AmpelAlertProtocol, filter_res: None | int = None) -> None:
 
 		alid = alert.id
 		if alid > self.alert_max:
