@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# File              : Ampel-alerts/ampel/dev/UnitTestAlertSupplier.py
-# License           : BSD-3-Clause
-# Author            : vb <vbrinnel@physik.hu-berlin.de>
-# Date              : 28.05.2020
-# Last Modified Date: 24.11.2021
-# Last Modified By  : vb <vbrinnel@physik.hu-berlin.de>
+# File:                Ampel-alerts/ampel/dev/UnitTestAlertSupplier.py
+# License:             BSD-3-Clause
+# Author:              valery brinnel <firstname.lastname@gmail.com>
+# Date:                28.05.2020
+# Last Modified Date:  24.11.2021
+# Last Modified By:    valery brinnel <firstname.lastname@gmail.com>
 
 from typing import List
 from ampel.protocol.AmpelAlertProtocol import AmpelAlertProtocol
@@ -26,7 +26,7 @@ class UnitTestAlertSupplier(AbsAlertSupplier):
 		)
 	)
 	"""
-	alerts: List[AmpelAlertProtocol]
+	alerts: list[AmpelAlertProtocol]
 
 	def __init__(self, **kwargs) -> None:
 		super().__init__(**kwargs)
