@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# File              : Ampel-alerts/ampel/alert/AlertConsumerError.py
-# License           : BSD-3-Clause
-# Author            : vb <vbrinnel@physik.hu-berlin.de>
-# Date              : 05.08.2021
-# Last Modified Date: 05.08.2021
-# Last Modified By  : vb <vbrinnel@physik.hu-berlin.de>
+# File:                Ampel-alerts/ampel/alert/AlertConsumerError.py
+# License:             BSD-3-Clause
+# Author:              valery brinnel <firstname.lastname@gmail.com>
+# Date:                05.08.2021
+# Last Modified Date:  05.08.2021
+# Last Modified By:    valery brinnel <firstname.lastname@gmail.com>
 
 import signal
 from enum import IntEnum
