@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# File              : Ampel-alerts/ampel/alert/load/TarballWalker.py
-# License           : BSD-3-Clause
-# Author            : jvs
-# Date              : 14.02.2019
-# Last Modified Date: 30.01.2020
-# Last Modified By  : vb <vbrinnel@physik.hu-berlin.de>
+# File:                Ampel-alerts/ampel/alert/load/TarballWalker.py
+# License:             BSD-3-Clause
+# Author:              jvs
+# Date:                14.02.2019
+# Last Modified Date:  30.01.2020
+# Last Modified By:    valery brinnel <firstname.lastname@gmail.com>
 
 import tarfile
 
